feat(WIP): reorganize code into web components
