+++ components/forms/add-private-key.js
@@ -1,0 +1,3 @@
+export const init = () => console.log(
+  'add-private-key.js not hidden',
+)
